+++ allowed-meta/index.test.ts
@@ -1,4 +1,4 @@
-import { ALLOWED_META } from '..'
+import { ALLOWED_META } from '../index.js'
 
 it('has allowed meta keys list', () => {
   for (let key of ALLOWED_META) {