+++ base-server/index.js
@@ -127,7 +127,7 @@ class BaseServer {
         try {
           resend = await processor.resend(ctx, action, meta)
         } catch (e) {
-          this.undo(meta, 'error')
+          this.undo(action, meta, 'error')
           this.emitter.emit('error', e, action, meta)
           this.finally(processor, ctx, action, meta)
           return
@@ -412,11 +412,11 @@ class BaseServer {
     })
   }
 
-  undo (meta, reason = 'error', extra = {}) {
+  undo (action, meta, reason = 'error', extra = {}) {
     let clientId = parseId(meta.id).clientId
-    let [action, undoMeta] = this.buildUndo(meta, reason, extra)
+    let [undoAction, undoMeta] = this.buildUndo(action, meta, reason, extra)
     undoMeta.clients = (undoMeta.clients || []).concat([clientId])
-    return this.log.add(action, undoMeta)
+    return this.log.add(undoAction, undoMeta)
   }
 
   debugError (error) {
@@ -432,26 +432,23 @@ class BaseServer {
   sendAction (action, meta) {
     let from = parseId(meta.id).clientId
     let clients = new Set([from])
-    let send = client => {
-      if (!clients.has(client.clientId)) {
-        clients.add(client.clientId)
-        client.node.onAdd(action, meta)
-      }
-    }
 
     if (meta.nodes) {
       for (let id of meta.nodes) {
         let client = this.nodeIds.get(id)
-        if (client && client.clientId !== from) {
-          send(client)
+        if (client) {
+          clients.add(client.clientId)
+          client.node.onAdd(action, meta)
         }
       }
     }
 
     if (meta.clients) {
       for (let id of meta.clients) {
-        if (this.clientIds.has(id) && id !== from) {
-          send(this.clientIds.get(id))
+        if (this.clientIds.has(id)) {
+          let client = this.clientIds.get(id)
+          clients.add(client.clientId)
+          client.node.onAdd(action, meta)
         }
       }
     }
@@ -461,8 +458,9 @@ class BaseServer {
         let users = this.userIds.get(userId)
         if (users) {
           for (let client of users) {
-            if (client.clientId !== from) {
-              send(client)
+            if (!clients.has(client.clientId)) {
+              clients.add(client.clientId)
+              client.node.onAdd(action, meta)
             }
           }
         }
@@ -516,7 +514,7 @@ class BaseServer {
     this.log.changeMeta(meta.id, { status: 'error' })
     this.reporter('unknownType', { type: action.type, actionId: meta.id })
     if (parseId(meta.id).userId !== 'server') {
-      this.undo(meta, 'unknownType')
+      this.undo(action, meta, 'unknownType')
     }
     this.debugActionError(meta, `Action with unknown type ${action.type}`)
   }
@@ -527,7 +525,7 @@ class BaseServer {
       actionId: meta.id,
       channel: action.channel
     })
-    this.undo(meta, 'wrongChannel')
+    this.undo(action, meta, 'wrongChannel')
     this.debugActionError(meta, `Wrong channel name ${action.channel}`)
   }
 
@@ -543,7 +541,7 @@ class BaseServer {
       this.markAsProcessed(meta)
     } catch (e) {
       this.log.changeMeta(meta.id, { status: 'error' })
-      this.undo(meta, 'error')
+      this.undo(action, meta, 'error')
       this.emitter.emit('error', e, action, meta)
     } finally {
       this.finally(processor, ctx, action, meta)
@@ -603,7 +601,7 @@ class BaseServer {
             return
           }
           if (!access) {
-            this.denyAction(meta)
+            this.denyAction(action, meta)
             return
           }
 
@@ -644,7 +642,7 @@ class BaseServer {
           this.markAsProcessed(meta)
         } catch (e) {
           this.emitter.emit('error', e, action, meta)
-          this.undo(meta, 'error')
+          this.undo(action, meta, 'error')
           if (subscribed) {
             this.unsubscribeAction(action, meta)
           }
@@ -681,9 +679,9 @@ class BaseServer {
     this.contexts.delete(action)
   }
 
-  denyAction (meta) {
+  denyAction (action, meta) {
     this.reporter('denied', { actionId: meta.id })
-    this.undo(meta, 'denied')
+    this.undo(action, meta, 'denied')
     this.debugActionError(meta, `Action "${meta.id}" was denied`)
   }
 
@@ -761,7 +759,7 @@ class BaseServer {
     }
   }
 
-  buildUndo (meta, reason, extra) {
+  buildUndo (action, meta, reason, extra) {
     let undoMeta = { status: 'processed' }
 
     if (meta.users) undoMeta.users = meta.users.slice(0)
@@ -770,8 +768,14 @@ class BaseServer {
     if (meta.reasons) undoMeta.reasons = meta.reasons.slice(0)
     if (meta.channels) undoMeta.channels = meta.channels.slice(0)
 
-    let action = { ...extra, type: 'logux/undo', id: meta.id, reason }
-    return [action, undoMeta]
+    let undoAction = {
+      ...extra,
+      type: 'logux/undo',
+      id: meta.id,
+      action,
+      reason
+    }
+    return [undoAction, undoMeta]
   }
 
   replaceResendShortcuts (meta) {