+++ filtered-node/index.test.ts
@@ -1,6 +1,6 @@
 import { ClientNode, TestPair, TestTime, TestLog } from '@logux/core'
 
-import FilteredNode from '../filtered-node'
+import FilteredNode from '../filtered-node/index.js'
 
 type Test = {
   client: ClientNode<{}, TestLog>