+++ index.d.ts
@@ -1,16 +1,16 @@
-import ServerClient from './server-client'
-import ALLOWED_META from './allowed-meta'
-import filterMeta from './filter-meta'
-import TestServer from './test-server'
-import TestClient from './test-client'
-import BaseServer from './base-server'
-import Context from './context'
-import Server from './server'
+import ServerClient from './server-client/index.js'
+import ALLOWED_META from './allowed-meta/index.js'
+import filterMeta from './filter-meta/index.js'
+import TestServer from './test-server/index.js'
+import TestClient from './test-client/index.js'
+import BaseServer from './base-server/index.js'
+import Context from './context/index.js'
+import Server from './server/index.js'
 
 export { Action } from '@logux/core'
 
-export { TestServerOptions } from './test-server'
-export { ServerOptions } from './server'
+export { TestServerOptions } from './test-server/index.js'
+export { ServerOptions } from './server/index.js'
 
 export {
   LoguxAction,
@@ -22,7 +22,7 @@ export {
   ServerMeta,
   BaseServerOptions,
   Logger
-} from './base-server'
+} from './base-server/index.js'
 
 export {
   ALLOWED_META,