+++ server/index.d.ts
@@ -1,4 +1,4 @@
-import BaseServer, { Logger, BaseServerOptions } from '../base-server'
+import BaseServer, { Logger, BaseServerOptions } from '../base-server/index.js'
 
 export type LoggerOptions = {
   /**
@@ -107,4 +107,16 @@ export default class Server<H extends object = {}> extends BaseServer<H> {
    * ```
    */
   options: ServerOptions
+
+  /**
+   * Load module creators and apply to the server. By default, it will load
+   * files from `modules/*`.
+   *
+   * ```js
+   * server.autoloadModules()
+   * ```
+   *
+   * @param files Pattern for module files.
+   */
+  autoloadModules (files: string[] | string): Promise<void>
 }